+++ src/Covenant/ASG.hs
@@ -39,8 +39,7 @@ module Covenant.ASG
         Builtin3,
         Builtin6,
         Lam,
-        Force,
-        Return
+        Force
       ),
     ValNodeInfo (Lit, App, Thunk, Cata, DataConstructor),
     ASGNode (..),
@@ -66,7 +65,6 @@ module Covenant.ASG
     builtin3,
     builtin6,
     force,
-    ret,
     lam,
     err,
     lit,
@@ -118,7 +116,6 @@ import Covenant.Internal.Rename
     renameValT,
     runRenameM,
     undoRename,
-    unsafeExperimentalRunRenameM,
   )
 import Covenant.Internal.Strategy (PlutusDataConstructor (PlutusB, PlutusConstr, PlutusI, PlutusList, PlutusMap))
 import Covenant.Internal.Term
@@ -131,8 +128,7 @@ import Covenant.Internal.Term
         Builtin3Internal,
         Builtin6Internal,
         ForceInternal,
-        LamInternal,
-        ReturnInternal
+        LamInternal
       ),
     CovenantTypeError
       ( ApplyCompType,
@@ -155,7 +151,6 @@ import Covenant.Internal.Term
         IntroFormWrongNumArgs,
         InvalidOpaqueField,
         LambdaResultsInNonReturn,
-        LambdaResultsInValType,
         NoSuchArgument,
         OutOfScopeTyVar,
         RenameArgumentFailed,
@@ -381,19 +376,13 @@ pattern Builtin6 f <- Builtin6Internal f
 pattern Force :: Ref -> CompNodeInfo
 pattern Force r <- ForceInternal r
 
--- | Produce the result of a computation.
---
--- @since 1.0.0
-pattern Return :: Ref -> CompNodeInfo
-pattern Return r <- ReturnInternal r
-
 -- | A lambda.
 --
--- @since 1.0.0
-pattern Lam :: Id -> CompNodeInfo
-pattern Lam i <- LamInternal i
+-- @since 1.2.0
+pattern Lam :: Ref -> CompNodeInfo
+pattern Lam r <- LamInternal r
 
-{-# COMPLETE Builtin1, Builtin2, Builtin3, Builtin6, Force, Return, Lam #-}
+{-# COMPLETE Builtin1, Builtin2, Builtin3, Builtin6, Force, Lam #-}
 
 -- | A compile-time literal of a flat builtin type.
 --
@@ -597,24 +586,6 @@ force r = do
     CompNodeType t -> throwError . ForceCompType $ t
     ErrorNodeType -> throwError ForceError
 
--- | Given the result of a function body (either a value or an error), construct
--- the return for it. Will fail if that reference aims at a computation node.
---
--- @since 1.0.0
-ret ::
-  forall (m :: Type -> Type).
-  (MonadHashCons Id ASGNode m, MonadError CovenantTypeError m) =>
-  Ref ->
-  m Id
-ret r = do
-  refT <- typeRef r
-  case refT of
-    ValNodeType t -> do
-      let t' = CompT count0 . CompTBody . NonEmpty.singleton $ t
-      refTo . ACompNode t' . ReturnInternal $ r
-    CompNodeType t -> throwError . ReturnCompType $ t
-    ErrorNodeType -> err
-
 -- | Given a desired type, and a computation which will construct a lambda body
 -- when executed (with the scope extended with the arguments the functions can
 -- expect), construct a lambda.
@@ -626,38 +597,38 @@ ret r = do
 -- \'bottom-up\', whereas function arguments (and their scopes) are necessarily
 -- top-down. Thus, we need to \'delay\' the construction of a lambda's body to
 -- ensure that proper scoped argument information can be given to it, hence why
--- the argument being passed is an @m Id@.
+-- the argument being passed is an @m Ref@.
 --
--- @since 1.0.0
+-- @since 1.2.0
 lam ::
   forall (m :: Type -> Type).
   (MonadHashCons Id ASGNode m, MonadError CovenantTypeError m, MonadReader ASGEnv m) =>
   CompT AbstractTy ->
-  m Id ->
+  m Ref ->
   m Id
-lam expectedT@(CompT cnt (CompTBody xs)) bodyComp = do
+lam expectedT@(CompT _ (CompTBody xs)) bodyComp = do
   let (args, resultT) = NonEmpty.unsnoc xs
-      cntW = view wordCount cnt
-  bodyId <- local (over (#scopeInfo % #argumentInfo) (Vector.cons (cntW, args))) bodyComp
-  bodyNode <- lookupRef bodyId
-  case bodyNode of
-    Nothing -> throwError . BrokenIdReference $ bodyId
-    -- This unifies with anything, so we're fine
-    Just AnError -> refTo . ACompNode expectedT . LamInternal $ bodyId
-    Just (ACompNode t specs) -> case specs of
-      ReturnInternal r -> do
-        rT <- typeRef r
-        case rT of
-          -- Note (Koz, 17/04/2025): I am not 100% sure about this, but I can't
-          -- see how anything else would make sense.
-          ValNodeType actualT ->
-            if resultT == actualT
-              then refTo . ACompNode expectedT . LamInternal $ bodyId
-              else throwError . WrongReturnType resultT $ actualT
-          ErrorNodeType -> throwError ReturnWrapsError -- Should be impossible
-          CompNodeType t' -> throwError . ReturnWrapsCompType $ t'
-      _ -> throwError . LambdaResultsInNonReturn $ t
-    Just (AValNode t _) -> throwError . LambdaResultsInValType $ t
+  bodyRef <- local (over (#scopeInfo % #argumentInfo) (Vector.cons args)) bodyComp
+  case bodyRef of
+    AnArg (Arg _ _ argTy) -> do
+      let argTy' = decDb argTy
+      if argTy' == resultT
+        then refTo . ACompNode expectedT . LamInternal $ bodyRef
+        else throwError . WrongReturnType resultT $ argTy'
+    AnId bodyId ->
+      lookupRef bodyId >>= \case
+        Nothing -> throwError . BrokenIdReference $ bodyId
+        -- This unifies with anything, so we're fine
+        Just AnError -> refTo . ACompNode expectedT . LamInternal . AnId $ bodyId
+        Just (AValNode ty _) -> do
+          let tyFixed = decDb ty
+          if tyFixed == resultT
+            then refTo . ACompNode expectedT . LamInternal . AnId $ bodyId
+            else throwError . WrongReturnType resultT $ tyFixed
+        Just (ACompNode t _) -> throwError $ LambdaResultsInCompType t
+  where
+    decDb :: ValT AbstractTy -> ValT AbstractTy
+    decDb = mapValT (\case Abstraction (BoundAt (S Z) argPos) -> Abstraction (BoundAt Z argPos); other -> other)
 
 -- | Construct the error node.
 --