+++ test/asg/Main.hs
@@ -11,8 +11,7 @@ import Covenant.ASG
     CompNodeInfo
       ( Builtin1,
         Builtin2,
-        Builtin3,
-        Return
+        Builtin3
       ),
     CovenantError (EmptyASG, TopLevelError, TopLevelValue, TypeError),
     CovenantTypeError
@@ -22,10 +21,8 @@ import Covenant.ASG
         ForceCompType,
         ForceError,
         ForceNonThunk,
-        LambdaResultsInValType,
         NoSuchArgument,
         OutOfScopeTyVar,
-        ReturnCompType,
         ThunkError,
         ThunkValType
       ),
@@ -43,15 +40,13 @@ import Covenant.ASG
     force,
     lam,
     lit,
-    nodeAt,
-    ret,
     runASGBuilder,
     thunk,
     topLevelNode,
   )
-import Covenant.Constant (AConstant (AUnit), typeConstant)
+import Covenant.Constant (typeConstant)
 import Covenant.DeBruijn (DeBruijn (S, Z))
-import Covenant.Index (Index, intIndex, ix0)
+import Covenant.Index (Index, intIndex, ix0, ix1)
 import Covenant.Prim
   ( typeOneArgFunc,
     typeThreeArgFunc,
@@ -81,7 +76,7 @@ import Test.QuickCheck
     (===),
   )
 import Test.Tasty (TestTree, adjustOption, defaultMain, testGroup)
-import Test.Tasty.HUnit (assertEqual, assertFailure, testCase)
+import Test.Tasty.HUnit (Assertion, assertEqual, assertFailure, testCase)
 import Test.Tasty.QuickCheck (QuickCheckTests, testProperty)
 
 main :: IO ()
@@ -95,9 +90,7 @@ main =
       testProperty "toplevel one-arg builtin compiles and has the right type" propTopLevelBuiltin1,
       testProperty "toplevel two-arg builtin compiles and has the right type" propTopLevelBuiltin2,
       testProperty "toplevel three-arg builtin compiles and has the right type" propTopLevelBuiltin3,
-      testProperty "toplevel return compiles and has the right type" propTopLevelReturn,
       testProperty "forcing a thunk has the same type as what the thunk wraps" propForceThunk,
-      testProperty "applying zero arguments to a return has the same type as what the return wraps" propApplyReturn,
       testProperty "forcing a computation type does not compile" propForceComp,
       testProperty "forcing a non-thunk value type does not compile" propForceNonThunk,
       testProperty "thunking a value type does not compile" propThunkValType,
@@ -106,10 +99,10 @@ main =
       testProperty "passing computations as arguments does not compile" propApplyComp,
       testProperty "requesting a non-existent argument does not compile" propNonExistentArg,
       testProperty "requesting an argument that exists compiles" propExistingArg,
-      testProperty "returning a computation from a lambda does not compile" propReturnComp,
-      testProperty "a lambda body having a value type does not compile" propLambdaValBody,
-      boundTyVarHappy,
-      boundTyVarShouldFail
+      testCase "db indices are well behaved (non-datatype case)" newLamTest1,
+      testCase "db indices are well behaved (datatype case)" newLamTest2,
+      testCase "calling down an in-scope tyvar works" boundTyVarHappy,
+      testCase "calling down an out-of-scope tyvar fails" boundTyVarShouldFail
     ]
   where
     moreTests :: QuickCheckTests -> QuickCheckTests
@@ -151,7 +144,7 @@ unitThunkError = do
 
 propTopLevelConstant :: Property
 propTopLevelConstant = forAllShrinkShow arbitrary shrink show $ \c ->
-  let builtUp = lit c
+  let builtUp = AnId <$> lit c
    in withCompilationFailure builtUp $ \case
         TopLevelValue _ t info -> case info of
           Lit c' ->
@@ -201,25 +194,6 @@ propTopLevelBuiltin3 = forAllShrinkShow arbitrary shrink show $ \bi3 ->
                 ]
             _ -> failUnexpectedCompNodeInfo info
 
-propTopLevelReturn :: Property
-propTopLevelReturn = forAllShrinkShow arbitrary shrink show $ \c ->
-  let builtUp = lit c >>= \i -> ret (AnId i)
-   in withCompilationSuccess builtUp $ \asg ->
-        withToplevelCompNode asg $ \t info ->
-          case info of
-            Return r -> withExpectedId r $ \i ->
-              withExpectedValNode i asg $ \t' info' ->
-                case info' of
-                  Lit c' ->
-                    let cT = typeConstant c
-                     in conjoin
-                          [ c' === c,
-                            cT === t',
-                            t === Comp0 (ReturnT cT)
-                          ]
-                  _ -> failUnexpectedValNodeInfo info'
-            _ -> failUnexpectedCompNodeInfo info
-
 -- We use builtins only for this test, but this should demonstrate the
 -- properties well enough
 propForceThunk :: Property
@@ -247,30 +221,15 @@ propForceThunk = forAllShrinkShow arbitrary shrink show $ \x ->
             force (AnId thunkI)
        in (comp, forceThunkComp)
 
--- As we can't build toplevel value ASGs, this has to be a bit roundabout
-propApplyReturn :: Property
-propApplyReturn = forAllShrinkShow arbitrary shrink show $ \c ->
-  let comp = do
-        i <- lit c
-        ret (AnId i)
-      applyReturnComp = do
-        i <- comp
-        applied <- app i Vector.empty mempty
-        ret (AnId applied)
-   in withCompilationSuccess comp $ \expectedASG ->
-        withCompilationSuccess applyReturnComp $ \applyReturnASG ->
-          withToplevelCompNode expectedASG $ \expectedT _ ->
-            withToplevelCompNode applyReturnASG $ \actualT _ ->
-              expectedT === actualT
-
 propForceComp :: Property
 propForceComp = forAllShrinkShow arbitrary shrink show $ \x ->
-  let comp = do
-        i <- case x of
-          Left bi1 -> builtin1 bi1
-          Right (Left bi2) -> builtin2 bi2
-          Right (Right bi3) -> builtin3 bi3
-        force (AnId i)
+  let comp =
+        AnId <$> do
+          i <- case x of
+            Left bi1 -> builtin1 bi1
+            Right (Left bi2) -> builtin2 bi2
+            Right (Right bi3) -> builtin3 bi3
+          force (AnId i)
       expectedT = case x of
         Left bi1 -> typeOneArgFunc bi1
         Right (Left bi2) -> typeTwoArgFunc bi2
@@ -282,9 +241,10 @@ propForceComp = forAllShrinkShow arbitrary shrink show $ \x ->
 
 propForceNonThunk :: Property
 propForceNonThunk = forAllShrinkShow arbitrary shrink show $ \c ->
-  let comp = do
-        i <- lit c
-        force (AnId i)
+  let comp =
+        AnId <$> do
+          i <- lit c
+          force (AnId i)
    in withCompilationFailure comp $ \case
         TypeError _ (ForceNonThunk actualT) -> typeConstant c === actualT
         TypeError _ err' -> failWrongTypeError err'
@@ -292,9 +252,10 @@ propForceNonThunk = forAllShrinkShow arbitrary shrink show $ \c ->
 
 propThunkValType :: Property
 propThunkValType = forAllShrinkShow arbitrary shrink show $ \c ->
-  let comp = do
-        i <- lit c
-        thunk i
+  let comp =
+        AnId <$> do
+          i <- lit c
+          thunk i
    in withCompilationFailure comp $ \case
         TypeError _ (ThunkValType actualT) -> typeConstant c === actualT
         TypeError _ err' -> failWrongTypeError err'
@@ -302,10 +263,11 @@ propThunkValType = forAllShrinkShow arbitrary shrink show $ \c ->
 
 propApplyToVal :: Property
 propApplyToVal = forAllShrinkShow arbitrary shrink show $ \c args ->
-  let comp = do
-        args' <- traverse (fmap AnId . lit) args
-        i <- lit c
-        app i args' mempty
+  let comp =
+        AnId <$> do
+          args' <- traverse (fmap AnId . lit) args
+          i <- lit c
+          app i args'
    in withCompilationFailure comp $ \case
         TypeError _ (ApplyToValType t) -> typeConstant c === t
         TypeError _ err' -> failWrongTypeError err'
@@ -313,10 +275,11 @@ propApplyToVal = forAllShrinkShow arbitrary shrink show $ \c args ->
 
 propApplyToError :: Property
 propApplyToError = forAllShrinkShow arbitrary shrink show $ \args ->
-  let comp = do
-        args' <- traverse (fmap AnId . lit) args
-        i <- err
-        app i args' mempty
+  let comp =
+        AnId <$> do
+          args' <- traverse (fmap AnId . lit) args
+          i <- err
+          app i args'
    in withCompilationFailure comp $ \case
         TypeError _ ApplyToError -> property True
         TypeError _ err' -> failWrongTypeError err'
@@ -332,13 +295,14 @@ propApplyComp = forAllShrinkShow arbitrary shrink show $ \f arg1 ->
         Right (Left bi2) -> typeTwoArgFunc bi2
         Right (Right bi3) -> typeThreeArgFunc bi3
 
-      comp = do
-        i <- builtin1 f
-        arg' <- case arg1 of
-          Left bi1 -> builtin1 bi1
-          Right (Left bi2) -> builtin2 bi2
-          Right (Right bi3) -> builtin3 bi3
-        app i (Vector.singleton . AnId $ arg') mempty
+      comp =
+        AnId <$> do
+          i <- builtin1 f
+          arg' <- case arg1 of
+            Left bi1 -> builtin1 bi1
+            Right (Left bi2) -> builtin2 bi2
+            Right (Right bi3) -> builtin3 bi3
+          app i (Vector.singleton . AnId $ arg')
    in withCompilationFailure comp $ \case
         TypeError _ (ApplyCompType actualT) -> t === actualT
         TypeError _ err' -> failWrongTypeError err'
@@ -346,7 +310,7 @@ propApplyComp = forAllShrinkShow arbitrary shrink show $ \f arg1 ->
 
 propNonExistentArg :: Property
 propNonExistentArg = forAllShrinkShow arbitrary shrink show $ \(db, index) ->
-  let comp = arg db index >>= \i -> ret (AnArg i)
+  let comp = arg db index >>= \i -> pure $ AnArg i
    in withCompilationFailure comp $ \case
         TypeError _ (NoSuchArgument db' index') -> conjoin [db === db', index === index']
         TypeError _ err' -> failWrongTypeError err'
@@ -359,7 +323,7 @@ propExistingArg :: Property
 propExistingArg = forAllShrinkShow gen shr show $ \(t, index) ->
   let comp = lam t $ do
         arg1 <- arg Z index
-        ret (AnArg arg1)
+        pure (AnArg arg1)
    in withCompilationSuccess comp $ \asg ->
         withToplevelCompNode asg $ \t' _ ->
           t' === t
@@ -399,38 +363,12 @@ propExistingArg = forAllShrinkShow gen shr show $ \(t, index) ->
                   Just (Concrete res') -> pure (Comp0 (ArgsAndResult (coerce args') res'), index)
            in shrinkOnIndex <|> shrinkOnArgs
 
-propReturnComp :: Property
-propReturnComp = forAllShrinkShow arbitrary shrink show $ \x ->
-  let t = case x of
-        Left bi1 -> typeOneArgFunc bi1
-        Right (Left bi2) -> typeTwoArgFunc bi2
-        Right (Right bi3) -> typeThreeArgFunc bi3
-      comp = do
-        i <- case x of
-          Left bi1 -> builtin1 bi1
-          Right (Left bi2) -> builtin2 bi2
-          Right (Right bi3) -> builtin3 bi3
-        ret (AnId i)
-   in withCompilationFailure comp $ \case
-        TypeError _ (ReturnCompType actualT) -> t === actualT
-        TypeError _ err' -> failWrongTypeError err'
-        err' -> failWrongError err'
-
-propLambdaValBody :: Property
-propLambdaValBody = forAllShrinkShow arbitrary shrink show $ \(Concrete t, c) ->
-  let resultT = typeConstant c
-      comp = lam (Comp0 (ArgsAndResult (Vector.singleton t) resultT)) $ lit c
-   in withCompilationFailure comp $ \case
-        TypeError _ (LambdaResultsInValType actualT) -> resultT === actualT
-        TypeError _ err' -> failWrongTypeError err'
-        err' -> failWrongError err'
-
-boundTyVarHappy :: TestTree
-boundTyVarHappy = testCase "boundTyVarHappy" . run $ do
+boundTyVarHappy :: Assertion
+boundTyVarHappy = run $ do
   lam lamTy $ do
     arg1 <- AnArg <$> arg Z ix0
     void $ boundTyVar Z ix0
-    ret arg1
+    pure arg1
   where
     lamTy :: CompT AbstractTy
     lamTy = Comp1 $ tyvar Z ix0 :--:> ReturnT (tyvar Z ix0)
@@ -440,8 +378,8 @@ boundTyVarHappy = testCase "boundTyVarHappy" . run $ do
       Left err' -> assertFailure . show $ err'
       Right {} -> pure ()
 
-boundTyVarShouldFail :: TestTree
-boundTyVarShouldFail = testCase "boundTyVarShouldFail" . run $ boundTyVar Z ix0
+boundTyVarShouldFail :: Assertion
+boundTyVarShouldFail = run $ boundTyVar Z ix0
   where
     run :: forall (a :: Type). (Show a) => ASGBuilder a -> IO ()
     run act = case runASGBuilder M.empty act of
@@ -452,91 +390,42 @@ boundTyVarShouldFail = testCase "boundTyVarShouldFail" . run $ boundTyVar Z ix0
       Left err' -> assertFailure $ "Expected an OutofScopeTyVar error, but got: " <> show err'
       Right x -> assertFailure $ "Expected boundTyVar to fail, but got: " <> show x
 
--- Intro form tests
-
-nothingIntro :: TestTree
-nothingIntro =
-  runIntroFormTest "nothing" expectNothingThunk $
-    dataConstructor "Maybe" "Nothing" mempty >>= typeIdVal
+-- TODO: better name
+newLamTest1 :: Assertion
+newLamTest1 = case runASGBuilder M.empty fn of
+  Left err' -> assertFailure (show err')
+  Right {} -> pure ()
   where
-    expectNothingThunk :: ValT AbstractTy
-    expectNothingThunk = ThunkT . Comp1 . ReturnT $ Datatype "Maybe" (Vector.fromList [tyvar Z ix0])
+    fn :: ASGBuilder Id
+    fn = lam expected $ do
+      f <- arg Z ix0 >>= force . AnArg
+      a <- AnArg <$> arg Z ix1
+      AnId <$> app f (Vector.singleton a)
 
-justConcreteIntro :: TestTree
-justConcreteIntro = runIntroFormTest "justConcreteIntro" expected $ do
-  argRef <- AnId <$> lit AUnit
-  dataConstructor "Maybe" "Just" (Vector.singleton argRef) >>= typeIdVal
+    expected :: CompT AbstractTy
+    expected =
+      Comp2 $
+        ThunkT (Comp0 $ tyvar (S Z) ix0 :--:> ReturnT (tyvar (S Z) ix1))
+          :--:> tyvar Z ix0
+          :--:> ReturnT (tyvar Z ix1)
+
+newLamTest2 :: Assertion
+newLamTest2 = case runASGBuilder tyAppTestDatatypes fn of
+  Left err' -> assertFailure (show err')
+  Right {} -> pure ()
   where
-    expected :: ValT AbstractTy
-    expected = ThunkT . Comp0 . ReturnT $ Datatype "Maybe" (Vector.singleton $ BuiltinFlat UnitT)
+    fn :: ASGBuilder Id
+    fn = lam expected $ do
+      f <- arg Z ix0 >>= force . AnArg
+      a <- AnArg <$> arg Z ix1
+      AnId <$> app f (Vector.singleton a)
 
-justRigidIntro :: TestTree
-justRigidIntro = runIntroFormTest "justRigidIntro" expected $ do
-  lamId <- lam lamTy $ do
-    arg1 <- AnArg <$> arg Z ix0
-    justRigid <- dataConstructor "Maybe" "Just" (Vector.singleton arg1)
-    ret (AnId justRigid)
-  lamThunked <- thunk lamId
-  typeIdVal lamThunked
-  where
-    lamTy :: CompT AbstractTy
-    lamTy = Comp1 $ tyvar Z ix0 :--:> ReturnT (ThunkT . Comp0 . ReturnT $ Datatype "Maybe" $ Vector.singleton (tyvar (S Z) ix0))
-
-    expected :: ValT AbstractTy
-    expected = ThunkT lamTy
-
-justNothingIntro :: TestTree
-justNothingIntro = runIntroFormTest "justNothingIntro" expectedThunk $ do
-  thunk <- lam expectedComp $ do
-    nothingThunk <- dataConstructor "Maybe" "Nothing" mempty
-    var <- boundTyVar Z ix0
-    nothingForced <- force (AnId nothingThunk)
-    nothingApplied <- app nothingForced mempty (Vector.singleton . wedgeLeft . Just $ var)
-    -- justNothing <- dataConstructor "Maybe" "Just" (Vector.singleton (AnId nothingApplied))
-    -- justNothingForced <-  force (AnId justNothing)
-    -- justNothingApplied <- app justNothingForced mempty (Vector.singleton Nowhere)
-    -- ret (AnId justNothingApplied)
-    -- justNothingInstantiated <- app justNothingForced mempty (Vector.singleton . wedgeLeft . Just $ var)
-    ret (AnId nothingApplied)
-  typeIdVal thunk
-  where
-    expectedComp :: CompT AbstractTy
-    expectedComp =
-      Comp1
-        . ReturnT
-        . Datatype "Maybe"
-        . Vector.singleton
-        -- . Datatype "Maybe"
-        -- \$ Vector.singleton
-        $ tyvar Z ix0
-
-    expectedThunk :: ValT AbstractTy
-    expectedThunk = ThunkT expectedComp
-
-runIntroFormTest :: String -> ValT AbstractTy -> DebugASGBuilder (ValT AbstractTy) -> TestTree
-runIntroFormTest nm expectedTy act = testCase nm $ case debugASGBuilder tyAppTestDatatypes act of
-  Left err -> assertFailure ("ASG Error: " <> show err)
-  Right actualTy -> assertEqual nm expectedTy actualTy
-
-debugUndoRename :: TestTree
-debugUndoRename = testCase "debugUndoRename" $ do
-  let expectNothingThunk :: ValT AbstractTy
-      expectNothingThunk = ThunkT . Comp1 . ReturnT $ Datatype "Maybe" (Vector.fromList [tyvar Z ix0])
-      input =
-        Comp0
-          . ReturnT
-          . ThunkT
-          . Comp1
-          . ReturnT
-          $ Datatype "Maybe" (Vector.fromList [tyvar Z ix0])
-  inputRenamed <-
-    either (error . show) pure
-      . runRenameM
-      . renameCompT
-      $ input
-  case checkApp tyAppTestDatatypes inputRenamed mempty of
-    Left err' -> assertFailure (show err')
-    Right actual -> assertEqual "debugUndoRename" expectNothingThunk (undoRename actual)
+    expected :: CompT AbstractTy
+    expected =
+      Comp2 $
+        ThunkT (Comp0 $ Datatype "Maybe" (Vector.singleton $ tyvar (S Z) ix0) :--:> ReturnT (tyvar (S Z) ix1))
+          :--:> Datatype "Maybe" (Vector.singleton $ tyvar Z ix0)
+          :--:> ReturnT (tyvar Z ix1)
 
 -- Helpers
 
@@ -546,7 +435,7 @@ failWrongTypeError err' = failWithCounterExample ("Unexpected type error: " <> s
 failWrongError :: CovenantError -> Property
 failWrongError err' = failWithCounterExample ("Unexpected error: " <> show err')
 
-withCompilationFailure :: ASGBuilder Id -> (CovenantError -> Property) -> Property
+withCompilationFailure :: ASGBuilder Ref -> (CovenantError -> Property) -> Property
 withCompilationFailure comp cb = case runASGBuilder M.empty comp of
   Left err' -> cb err'
   Right asg -> failWithCounterExample ("Unexpected success: " <> show asg)
@@ -572,6 +461,7 @@ failUnexpectedValNodeInfo :: ValNodeInfo -> Property
 failUnexpectedValNodeInfo info =
   failWithCounterExample ("Unexpected ValNodeInfo: " <> show info)
 
+{- NOTE: Not 100% sure I won't need these
 withExpectedId :: Ref -> (Id -> Property) -> Property
 withExpectedId r cb = case r of
   AnId i -> cb i
@@ -581,3 +471,4 @@ withExpectedValNode :: Id -> ASG -> (ValT AbstractTy -> ValNodeInfo -> Property)
 withExpectedValNode i asg cb = case nodeAt i asg of
   AValNode t info -> cb t info
   node -> failWithCounterExample ("Unexpected node: " <> show node)
+--}