+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## UNRELEASED
 
 * Helper function for safely retrieving an in-scope type variable when constructing the ASG  
+* Removed Return nodes, associated pattern synonyms, and the `ret` ASGBuilder function
 
 ## 1.1.0 -- 2025-07-11